fix: restore auth and admin page rendering
