+++ app/(protected)/audit-students/page.tsx
@@ -1,9 +1,11 @@
+import { Suspense } from "react";
 import { redirect } from "next/navigation";
 import { auth } from "@/lib/auth";
 import StudentAuditPage from "@/components/audit/student-audit-page";
 import { Metadata } from "next";
 
 export const runtime = "nodejs";
+export const dynamic = "force-dynamic";
 
 export const metadata: Metadata = {
   title: "Student Audit",
@@ -21,5 +23,15 @@ export default async function AuditStudentsPage() {
     redirect("/dashboard");
   }
 
-  return <StudentAuditPage />;
+  return (
+    <Suspense
+      fallback={
+        <div className="min-h-screen bg-terminal-dark flex items-center justify-center p-4">
+          <div className="h-6 w-6 border-2 border-terminal-green border-t-transparent rounded-full animate-spin" />
+        </div>
+      }
+    >
+      <StudentAuditPage />
+    </Suspense>
+  );
 }
