+++ app/api/auth/reset-password/route.ts
@@ -2,38 +2,17 @@ import { NextRequest, NextResponse } from "next/server";
 import { prisma } from "@/lib/prisma";
 import { hashPassword } from "@/lib/auth";
 import { createAuditLog } from "@/lib/audit";
+import { verifyTurnstileToken } from "@/lib/security";
 
 export async function POST(request: NextRequest) {
   try {
     const { token, password, turnstileToken } = await request.json();
 
-    // Verify Turnstile token
-    if (!turnstileToken) {
+    const turnstileVerification = await verifyTurnstileToken(turnstileToken);
+
+    if (!turnstileVerification.ok) {
       return NextResponse.json(
-        { error: "CAPTCHA verification required" },
-        { status: 400 },
-      );
-    }
-
-    const turnstileResponse = await fetch(
-      "https://challenges.cloudflare.com/turnstile/v0/siteverify",
-      {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/x-www-form-urlencoded",
-        },
-        body: new URLSearchParams({
-          secret: process.env.TURNSTILE_SECRET_KEY!,
-          response: turnstileToken,
-        }),
-      },
-    );
-
-    const turnstileResult = await turnstileResponse.json();
-
-    if (!turnstileResult.success) {
-      return NextResponse.json(
-        { error: "CAPTCHA verification failed" },
+        { error: turnstileVerification.error },
         { status: 400 },
       );
     }
