--- app/metadata.ts
@@ -1,42 +1,0 @@
-export const metadata = {
-  title: {
-    default: "Codezela Career Accelerator - LMS",
-    template: "%s | Codezela Career Accelerator",
-  },
-  description:
-    "Codezela Career Accelerator Learning Management System - Empowering future tech leaders with comprehensive programming courses, career development, and hands-on learning experiences.",
-  keywords: [
-    "Codezela",
-    "Career Accelerator",
-    "LMS",
-    "Learning Management System",
-    "Programming Courses",
-    "Tech Education",
-    "Online Learning",
-    "Career Development",
-    "Skill Building",
-    "Tech Training",
-  ],
-  authors: [{ name: "Codezela" }],
-  creator: "Codezela",
-  publisher: "Codezela",
-  applicationName: "Codezela Career Accelerator LMS",
-  robots: "noindex, nofollow",
-  themeColor: [
-    { media: "(prefers-color-scheme: light)", color: "#f0f4f0" },
-    { media: "(prefers-color-scheme: dark)", color: "#0a0f0a" },
-  ],
-  openGraph: {
-    type: "website",
-    siteName: "Codezela Career Accelerator",
-    title: "Codezela Career Accelerator - LMS",
-    description:
-      "Empowering future tech leaders with comprehensive programming courses and career development.",
-  },
-};
-
-export const viewport = {
-  width: "device-width",
-  initialScale: 1,
-  maximumScale: 1,
-};
